+++ lib/icon_content.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+const iconSize = 80.0;
+const sizedBoxHeight = 15.0;
+const labelTextStyle = TextStyle(fontSize: 18.0, color: Color(0xFF8D8E98));
+
+class IconContent extends StatelessWidget {
+  IconContent({this.icon, this.label});
+
+  final IconData icon;
+  final String label;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: <Widget>[
+        Icon(
+          icon,
+          size: iconSize,
+        ),
+        SizedBox(
+          height: sizedBoxHeight,
+        ),
+        Text(
+          label,
+          style: labelTextStyle,
+        )
+      ],
+    );
+  }
+}
